Remove stack children type: - ASLayoutable requires mutable properties that are used when attached to a stack layout. - Thus, ASLayoutable objects (including ASDisplayNode) can be injected into stack layout directly. - ASStackLayoutNodeChild no longer needed. - Tests and Kitten sample updated.
